+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - feat(camera): make camera service constantly report its input to the redux bus when viewfinder is open
 - feat(assistant): implement tool call for providing camera-feed/display-content-feed to the llm
 - refactor(assistant): improve the mechanism of applying `is_listening` changes to the input transport of the assistant client
+- feat(keypad): add hold and unhold actions and use them to run assistant anywhere by holding the home key in
 
 ## Version 1.6.0